% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{as_tibble.alias.matrix}
\alias{as_tibble.alias.matrix}
\title{Convert an alias matrix to a tibble}
\usage{
\method{as_tibble}{alias.matrix}(
  am,
  ...,
  .rows = NULL,
  .name_repair = "check_unique",
  rownames = NULL,
  compact = TRUE
)
}
\arguments{
\item{am}{the alias matrix object}

\item{...}{additional arguments to \code{as_tibble}}

\item{.rows}{The number of rows, useful to create a 0-column tibble or just as an additional check (see \code{\link[tibble:as_tibble]{tibble::as_tibble()}})}

\item{.name_repair}{Treatment of problematic column names (see \code{\link[tibble:as_tibble]{tibble::as_tibble()}})}

\item{rownames}{How to treat existing row names of a data frame or matrix (see \code{\link[tibble:as_tibble]{tibble::as_tibble()}}}

\item{compact}{a logical: if TRUE, it reports all possible effects
combinations, even those with no alias.}
}
\value{
a tibble
}
\description{
Given an alias matrix, this function returns a tidy tibble of the
alias structures, with the added \code{generator} column containing the generator
(i.e. right-hand side) of the defining relationship that generates each
alias.
}
\examples{
tibble::as_tibble(fp_alias_matrix(~A*B*C, ~B*C*D))
}
