% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccd_experiment_yield}
\alias{ccd_experiment_yield}
\title{Central Composite Design Experiment Yields}
\format{
A list with three vectors:
\itemize{
\item \code{base}: the yield for a 2^2 factorial design, replicated 3 times
\item \code{center}: the yield for the center points, replicated 4 times
\item \code{axial}: the yield for the axial points, replicated 2 times
}
}
\usage{
ccd_experiment_yield
}
\description{
Yield data for a two factor CCD experiment
}
\keyword{datasets}
