% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{chauvenet}
\alias{chauvenet}
\title{Chauvenet's criterion}
\usage{
chauvenet(x, threshold = 0.5)
}
\arguments{
\item{x}{the sample vector}

\item{threshold}{the threshold for the frequency of the suspect outlier}
}
\value{
a list with the following components:
\itemize{
\item{s0}{the maximum difference}
\item{index}{the index of the suspect outlier}
\item{value}{the value of the suspect outlier}
\item{reject}{a logical value indicating whether the suspect outlier should be rejected}
}
}
\description{
Applies the Chauvenet's criterion to a sample, identifying a possible
outlier.
}
\examples{
x <- rnorm(100)
chauvenet(x)
chauvenet(x, threshold=0.1)
}
