% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cotton}
\alias{cotton}
\title{Cotton yarn experiment data}
\format{
A data frame with 25 rows and 3 columns. Columns represent:
\itemize{
\item \code{Run}: run order
\item \code{Cotton}: cotton content in mass percentage
\item \code{Strength}: yarn tensile strength in N
}
}
\usage{
cotton
}
\description{
Yarn tensile strength in a completely randomized experiment with 5 different
levels of cotton fiber.
}
\references{
Douglas C. Montgomery, "Design and Analysis of Experiments", 8th edition, Wiley, 2019
}
\keyword{datasets}
