% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daniel_plot}
\alias{daniel_plot}
\title{Daniel's plot}
\usage{
daniel_plot(model, alpha = 0.5, xlim = c(-3, 3))
}
\arguments{
\item{model}{a linear model}

\item{alpha}{the transparency of the horizontal lines}

\item{xlim}{the limits of the x-axis}
}
\value{
a QQ plot with the effects of the model
}
\description{
Daniel's plot
}
\examples{
daniel_plot(lm(mpg ~ wt, data=mtcars))
}
