% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{design_matrix}
\alias{design_matrix}
\title{Generate the design matrix for a factorial experiment}
\usage{
design_matrix(n_factors, rep = 1, levels = c(-1, 1))
}
\arguments{
\item{n_factors}{the number of factors (a scalar)}

\item{rep}{the number of replications (a scalar)}

\item{levels}{the levels of the factors (a vector)}
}
\value{
a design matrix (as a tibble). The design matrix also has the
\code{StdOrder} and \code{RunOrder} columns, which are the standard order and the
random order of the runs, respectively. Also, a NA column \code{Y} is added for
the response variable.
}
\description{
The function produces a design matrix for a factorial experiment. The
function takes the number of factors, the number of replications, and the
levels of the factors as arguments. The function returns a tibble with the
design matrix. The design matrix also has the \code{StdOrder} and \code{RunOrder}
columns, which are the standard order and the random order of the runs,
respectively. Also, a NA column \code{Y} is added for the response variable.
}
\examples{
# A 3^3 factorial plan, teplicated 2 times:
design_matrix(3, rep=2, levels=-1:1)
}
