% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{examples_url}
\alias{examples_url}
\title{examples_url
provides the URL for the desired example data}
\usage{
examples_url(example)
}
\arguments{
\item{example}{data file name}
}
\value{
the full URL for the desired example
}
\description{
examples_url
provides the URL for the desired example data
}
\examples{
examples_url("train.csv")
}
