% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filtration}
\alias{filtration}
\title{Filtration data}
\format{
Factors are:
\itemize{
\item A: Temperature
\item B: Pressure
\item C: Concentration of solid phase
\item D: Agitation speed
}

The yield is in column Y and represents the filtration speed
}
\usage{
filtration
}
\description{
Non-replicated factorial plan for a slurry filtration process.
}
\keyword{datasets}
