% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_alias}
\alias{fp_alias}
\title{Tabulate all the Aliases from a Defining Relationship}
\usage{
fp_alias(arg)
}
\arguments{
\item{arg}{A formula for the defining relationship}
}
\value{
A matrix of aliases
}
\description{
Tabulates all the aliases from a defining relationship.
}
\examples{
fp_alias(~A*B*C*D)
}
\seealso{
\code{\link[=fp_defrel]{fp_defrel()}}
}
