% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_alias_matrix}
\alias{fp_alias_matrix}
\title{Build the alias matrix}
\usage{
fp_alias_matrix(...)
}
\arguments{
\item{...}{one or more formulas, or a single list of formulas}
}
\value{
a square matrix
}
\description{
Given a list of formulas (defining relationships), this function returns
a matrix of all possible aliases.
}
\examples{
fp_alias_matrix(~A*B*C, ~B*C*D)
}
