% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_all_drs}
\alias{fp_all_drs}
\title{Return a list of all defining relationships}
\usage{
fp_all_drs(...)
}
\arguments{
\item{...}{formulas, or a single list of formulas.}
}
\value{
a list of formulas.
}
\description{
Given two or more independent refining relationships, represented as one
side formulas,, this function returns a list of all possible defining
relationships, including the dependent ones.
}
\details{
Defining relationships are represented as one side formulas, e.g. $I=ABC$
becomes \code{~A*B*C}.
}
\examples{
fp_all_drs(~A*B*C, ~B*C*D)
}
