% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_augment_center}
\alias{fp_augment_center}
\title{Augment to a central composite design}
\usage{
fp_augment_center(dm, rep = 5)
}
\arguments{
\item{dm}{A factorial plan table}

\item{rep}{The number of replications}
}
\value{
A central composite design
}
\description{
Add the central points to an existing $2^n$ factorial plan.
}
\examples{
fp_design_matrix(3) \%>\%
 fp_augment_center()
}
