% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_defrel}
\alias{fp_defrel}
\title{Factorial Plan Defining Relationship}
\usage{
fp_defrel(arg)
}
\arguments{
\item{arg}{If it is a formula, it is returned verbatim. If it is a number, a
formula is built with the number of factors. If it is neither a formula nor a
number, an error is thrown.}
}
\value{
A formula
}
\description{
Builds a formula from a number of factors
}
\examples{
# Defining relationships with three factors
fp_defrel(3)

# Defining relationship I=ABC
fp_defrel(~A*B*C)
}
