% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_design_matrix}
\alias{fp_design_matrix}
\title{Factorial Plan Design Matrix}
\usage{
fp_design_matrix(arg, rep = 1, levels = c(-1, 1))
}
\arguments{
\item{arg}{Either a formula or a number of factors}

\item{rep}{Number of replications}

\item{levels}{Levels of the factors}
}
\value{
A design matrix: a subclass of a tibble of class \code{factorial.plan}
}
\description{
Factorial Plan Design Matrix
}
\examples{
fp_design_matrix(3, rep=2, levels=c("-", "+"))
}
