% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_gen2alias}
\alias{fp_gen2alias}
\title{Given a generator, find the alias}
\usage{
fp_gen2alias(generator, effect)
}
\arguments{
\item{generator}{a generator, in the form of \code{ABCD...}.}

\item{effect}{an effect, in the form of \code{BD...}.}
}
\value{
An effect (string).
}
\description{
Given a generator and an effect, this function returns the alias.
}
\details{
Generators and aliases are strings of capital letters.
}
\examples{
fp_gen2alias("ABCD", "BD")
}
