% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_has_alias}
\alias{fp_has_alias}
\title{Check for Factor Aliases}
\usage{
fp_has_alias(arg, A, B)
}
\arguments{
\item{arg}{A formula for the defining relationship}

\item{A}{a string representing an effect (e.g. \code{"AB"})}

\item{B}{a string representing an effect (e.g. \code{"CD"})}
}
\value{
A logical value
}
\description{
Checks if two factors are aliased in a formula. This function is
moslty used internally to build the alias matrix.
}
\examples{
fp_has_alias(~A*B*C*D, "AB", "CD")
}
\seealso{
\code{\link[=fp_defrel]{fp_defrel()}} \code{\link[=fp_alias]{fp_alias()}} \code{\link[=fp_alias_list]{fp_alias_list()}}
}
