% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_info}
\alias{fp_info}
\title{Factorial plan info}
\usage{
fp_info(x, file = "", comment = "")
}
\arguments{
\item{x}{the factorial plan.}

\item{file}{the file to write the information to. Use console if empty.}

\item{comment}{a comment mark to add before each line of the information.}
}
\value{
No return value, just prints the fp information.
}
\description{
Print information about the factorial plan.
}
\examples{
fp_design_matrix(3, rep=2) \%>\%
  fp_info()
}
