% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_merge_drs}
\alias{fp_merge_drs}
\title{Return a merged defining relationship}
\usage{
fp_merge_drs(f1, ...)
}
\arguments{
\item{f1}{a formula.}

\item{...}{other formulas.}
}
\value{
a formula.
}
\description{
This function, given one or more independent refining relationships, returns
the most complete relationship, i.e. that which includes all the factors.
}
\details{
Defining relationships are represented as one side formulas, e.g. $I=ABC$
becomes \code{~A*B*C}.
}
\examples{
fp_merge_drs(~A*B*C, ~B*C*D)
}
