% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_read_csv}
\alias{fp_read_csv}
\title{Load a design matrix from a CSV file}
\usage{
fp_read_csv(dm, file, type = c(1, 2), yield = "Y", comment = "#")
}
\arguments{
\item{dm}{the design matrix.}

\item{file}{the file to read the design matrix from.}

\item{type}{the CSV version (1 or 2).}

\item{yield}{the yield column name.}

\item{comment}{the comment mark.}
}
\value{
the design matrix with the new values.
}
\description{
Load from a CSV file the design matrix that has previously been saved with
\code{fp_write_csv()}. It is an error if the loaded data frame has different
dimensions or column names than the original design matrix.
}
\details{
Note that the design matrix is sorted by the \code{StdOrder} column after loading.
}
\seealso{
\code{\link[=fp_write_csv]{fp_write_csv()}}
}
