% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_treatments}
\alias{fp_treatments}
\title{Factorial Plan List of Treatments}
\usage{
fp_treatments(arg)
}
\arguments{
\item{arg}{Either a formula or a number of factors}
}
\value{
A list of treatments
}
\description{
Builds a list of treatments from a formula, or from a number of factors.
}
\examples{
fp_treatments(3)
}
