% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_write_csv}
\alias{fp_write_csv}
\title{Save a design matrix to a CSV file}
\usage{
fp_write_csv(dm, file, comment = "# ", timestamp = TRUE, type = c(1, 2), ...)
}
\arguments{
\item{dm}{the design matrix.}

\item{file}{the file to write the design matrix to.}

\item{comment}{a comment mark to add before each line of the information.}

\item{timestamp}{whether to add a timestamp to the file.}

\item{type}{the CSV version (1 or 2).}

\item{...}{other parameters passed to write_csv().}
}
\value{
Invisibly return the design matrix, unchanged, for further piping.
}
\description{
Writes the design matrix to a CSV file, with a timestamp and comment lines.
}
\details{
Note that the design matrix is saved in the same order of the \code{RunOrder}
column, i.e. random.
}
