% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ggTukey}
\alias{ggTukey}
\title{ggTukey}
\usage{
ggTukey(obj, ...)
}
\arguments{
\item{obj}{an object to plot: either a TukeyHSD object or a data frame}

\item{...}{Other parameters passed to the specialized functions}
}
\value{
a GGPlot2 object
}
\description{
This is a generic function for plotting Tukey's HSD test via GGplot2.
}
