% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pareto_chart.default}
\alias{pareto_chart.default}
\title{Pareto's chart}
\usage{
\method{pareto_chart}{default}(data, labels, values)
}
\arguments{
\item{data}{a data frame}

\item{labels}{the column with the labels of the data frame}

\item{values}{the column with the values of the data frame}
}
\value{
a Pareto chart of the data frame

Invisibly returns a data frame with the absolute values of the
data frame, their sign, and the cumulative value.
}
\description{
Create a Pareto chart for a data frame.
}
\examples{
pareto_chart(mtcars)
}
