% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pareto_chart}
\alias{pareto_chart}
\title{Pareto's chart}
\usage{
pareto_chart(...)
}
\arguments{
\item{model}{a model}
}
\value{
a Pareto chart of the effects of the model
}
\description{
This is a generic function for Pareto's chart.
}
\examples{
pareto_chart(lm(mpg ~ wt + hp, data=mtcars))
}
