% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{plot.alias.matrix}
\alias{plot.alias.matrix}
\title{Plot the alias matrix}
\usage{
\method{plot}{alias.matrix}(x, ..., compact = TRUE)
}
\arguments{
\item{x}{an alias matrix.}

\item{...}{additional arguments to \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}.}

\item{compact}{logical, if TRUE only positive aliases are shown, omitting
empty rows and columns.}
}
\value{
a ggplot object.
}
\description{
Produces a tile plot of the alias matrix.
}
\examples{
fp_alias_matrix(~A*B*C, ~B*C*D) \%>\%
  plot()
}
