% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adas.utils-package.R
\docType{package}
\name{adas.utils-package}
\alias{adas.utils}
\alias{adas.utils-package}
\title{adas.utils: Utility functions for the ADAS course, University of Trento, Italy}
\description{
Utility functions for the ADAS course, University of Trento, Italy. The package contains functions to deal with Design of Experiments (DoE) factorial plans, with the addition of some utility function to perform some statistical analyses.
}
\author{
\strong{Maintainer}: Paolo Bosetti \email{paolo.bosetti@unitn.it}

}
\keyword{internal}
