% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adas.utils-package.R
\docType{package}
\name{adas.utils-package}
\alias{adas.utils}
\alias{adas.utils-package}
\title{adas.utils: Utility functions for the ADAS course, University of Trento, Italy}
\description{
Utility functions for the ADAS course, University of Trento, Italy
}
\author{
\strong{Maintainer}: Paolo Bosetti \email{paolo.bosetti@unitn.it}

}
\keyword{internal}
