% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{daniel_plot_hn}
\alias{daniel_plot_hn}
\title{Daniel's plot (half-normal)}
\usage{
daniel_plot_hn(model, ...)
}
\arguments{
\item{model}{a linear model}

\item{...}{further arguments to \code{\link[gghalfnorm:gghalfnorm]{gghalfnorm::gghalfnorm()}}}
}
\value{
a half-normal plot (GGPlot2 object) with the effects of the model
}
\description{
Given a non-replicated model of a factorial plan, this function provides a
half-normal plot of the effects of the model, labeling the main n effects.
}
\examples{
daniel_plot_hn(lm(Y~A*B*C*D, data=filtration))
}
\seealso{
\code{\link[gghalfnorm:gghalfnorm]{gghalfnorm::gghalfnorm()}}
}
