% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{examples_url}
\alias{examples_url}
\title{Examples URL}
\usage{
examples_url(example)
}
\arguments{
\item{example}{data file name}
}
\value{
the full URL for the desired example
}
\description{
Provides the URL for the desired example data, so that it can be more easily
downloaded.
}
\examples{
examples_url("battery.dat") |> read.table(header=TRUE)
}
