% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_alias}
\alias{fp_alias}
\title{Build an alias table from a formula}
\usage{
fp_alias(arg)
}
\arguments{
\item{arg}{A formula or a number of factors.}
}
\value{
A matrix of logical values.
}
\description{
Given a Defining relationship or a number of factors, this function builds
a matrix of aliases.
}
\details{
Defining relationships are represented as one side formulas, e.g. $I=ABC$
becomes \code{~A*B*C}.
}
\examples{
fp_alias(~A*B*C*D)
fp_alias(3)
}
