% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_alias_list}
\alias{fp_alias_list}
\title{List All Alias for a Fractional Factorial Plan}
\usage{
fp_alias_list(arg)
}
\arguments{
\item{arg}{A formula for the defining relationship, or the number of factors.}
}
\value{
a list of aliases (as formulas).
}
\description{
Given a defining relationship, as a one side firmula,, this function lists
all the aliases for a fractional factorial plan.
}
\details{
Defining relationships are represented as one side formulas, e.g. $I=ABC$
becomes \code{~A*B*C}.
}
\examples{
fp_alias_list(~A*B*C*D)
}
