% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_effect_names}
\alias{fp_effect_names}
\title{Factorial Plan effect names from a formula}
\usage{
fp_effect_names(arg)
}
\arguments{
\item{arg}{A formula.}
}
\value{
An ordered factor with the effect names.
}
\description{
Returns the effect names from a formula, according to Yates' convention.
}
\details{
Defining relationships are represented as one side formulas, e.g. $I=ABC$
becomes \code{~A*B*C}.
}
\examples{
fp_effect_names(~A*B*C)
}
