% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_plan.R
\name{fp_treatments}
\alias{fp_treatments}
\title{Factorial Plan List of Treatments}
\usage{
fp_treatments(arg)
}
\arguments{
\item{arg}{Either a formula or a number of factors. If it is a formula, the
factors are extracted from it. If it is a number, the factors are the first
\code{n} capital letters.}
}
\value{
A list of treatments (character vector).
}
\description{
Builds a list of treatments from a formula, or from a number of factors.
}
\details{
Defining relationships are represented as one side formulas, e.g. $I=ABC$
becomes \code{~A*B*C}.
}
\examples{
fp_treatments(3)
}
