% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pareto_chart.lm}
\alias{pareto_chart.lm}
\title{Pareto's chart}
\usage{
\method{pareto_chart}{lm}(obj, ...)
}
\arguments{
\item{obj}{a linear model}

\item{...}{further parameters (currently unused)}
}
\value{
a Pareto chart (GGPlot2 object) of the effects of the model

Invisibly returns a data frame with the absolute effects of the
model, their sign, and the cumulative effect.
}
\description{
Creates a Pareto chart for the effects of a linear model.
}
\examples{
pareto_chart(lm(Y~A*B*C*D, data=filtration))
}
