% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pareto_chart}
\alias{pareto_chart}
\title{Pareto's chart}
\usage{
pareto_chart(model)
}
\arguments{
\item{model}{a linear model}
}
\value{
a Pareto chart of the effects of the model
}
\description{
Pareto's chart
}
\examples{
pareto_chart(lm(mpg ~ wt + hp, data=mtcars))
}
